Switches timer get API to uint64
Use Ogre::uint64 in Timer get functions instead of unsigned long long to ensure 64-bit size. Some of the get functions require 64-bit to not overflow. All return uint64 for consistency.

+++ OgreMain/include/OgreFrameStats.h
@@ -48,7 +48,7 @@ namespace Ogre
         int mNextFrame;
         unsigned long mBestFrameTime;
         unsigned long mWorstFrameTime;
-        unsigned long long mLastTime;
+        uint64 mLastTime;
         unsigned long mFrameTimes[OGRE_FRAME_STATS_SAMPLES];
         size_t mFramesSampled;
 
@@ -88,7 +88,7 @@ namespace Ogre
         }
 
         /// Adds a new measured time, in *microseconds*
-        void addSample( unsigned long long timeMs )
+        void addSample( uint64 timeMs )
         {
             unsigned long frameTimeMs = static_cast<unsigned long>(timeMs - mLastTime);
             mFrameTimes[mNextFrame]  = frameTimeMs;
@@ -101,7 +101,7 @@ namespace Ogre
             mLastTime = timeMs;
         }
 
-        void reset( unsigned long long timeMs )
+        void reset( uint64 timeMs )
         {
             mNextFrame   = 0;
             mBestFrameTime  = std::numeric_limits<unsigned long>::max();
