+++ OgreMain/include/iOS/OgreTimerImp.h
@@ -63,16 +63,16 @@ namespace Ogre
         void reset();
 
         /** Returns milliseconds since initialisation or last reset */
-        unsigned long long getMilliseconds();
+        uint64 getMilliseconds();
 
         /** Returns microseconds since initialisation or last reset */
-        unsigned long long getMicroseconds();
+        uint64 getMicroseconds();
 
         /** Returns milliseconds since initialisation or last reset, only CPU time measured */  
-        unsigned long long getMillisecondsCPU();
+        uint64 getMillisecondsCPU();
 
         /** Returns microseconds since initialisation or last reset, only CPU time measured */  
-        unsigned long long getMicrosecondsCPU();
+        uint64 getMicrosecondsCPU();
     };
 }
 #endif
