+++ OgreMain/src/OSX/OgreTimer.cpp
@@ -49,7 +49,7 @@ void Timer::reset()
 }
 
 //--------------------------------------------------------------------------------//
-unsigned long long Timer::getMilliseconds()
+uint64 Timer::getMilliseconds()
 {
     struct timeval now;
     gettimeofday(&now, NULL);
@@ -57,7 +57,7 @@ unsigned long long Timer::getMilliseconds()
 }
 
 //--------------------------------------------------------------------------------//
-unsigned long long Timer::getMicroseconds()
+uint64 Timer::getMicroseconds()
 {
     struct timeval now;
     gettimeofday(&now, NULL);
@@ -65,15 +65,15 @@ unsigned long long Timer::getMicroseconds()
 }
 
 //-- Common Across All Timers ----------------------------------------------------//
-unsigned long long Timer::getMillisecondsCPU()
+uint64 Timer::getMillisecondsCPU()
 {
     clock_t newClock = clock();
-    return (unsigned long long )((float)(newClock-zeroClock) / ((float)CLOCKS_PER_SEC/1000.0)) ;
+    return (uint64 )((float)(newClock-zeroClock) / ((float)CLOCKS_PER_SEC/1000.0)) ;
 }
 
 //-- Common Across All Timers ----------------------------------------------------//
-unsigned long long Timer::getMicrosecondsCPU()
+uint64 Timer::getMicrosecondsCPU()
 {
     clock_t newClock = clock();
-    return (unsigned long long)((float)(newClock-zeroClock) / ((float)CLOCKS_PER_SEC/1000000.0)) ;
+    return (uint64)((float)(newClock-zeroClock) / ((float)CLOCKS_PER_SEC/1000000.0)) ;
 }
